Correctly handle deeply nested scopes
When large batch DOM updates are made the document mutation observer that
handles rescoping things will fire with only the root most node of the newly
attached tree. This is expected behavior of mutation observer. As a result it's
possible that multiple shadow roots will be appended in a single tick of the
observer with unscoped children. Previously every unscoped child would be scoped
to the scope of the root most scope. However this can be incorrect when multiple
shadow boundaries have been appended in a tree. In order to prevent incorrectly
scoping children nodes before scoping them their root and host need to be
identified. The correct scope will then be applied instead of assuming that it's
the scope of the newly added node.

Even when the `newScope` is in fact different from the `currentScope` it's possible
that unscoped children could be added that need a new and different scope from
the root node.